requirements.txt
python-telegram-bot
pytesseract
fuzzywuzzy[speedup]